backend/requirements.txt
# Django Framework
Django==5.2.3
djangorestframework==3.15.0
djangorestframework-simplejwt==5.2.2
django-cors-headers==4.7.0

# Database
pymongo==4.13.0
dnspython==2.7.0

# HTTP Requests
requests==2.32.3

# Web Scraping
beautifulsoup4==4.12.3
lxml==5.1.0

# AI/ML Libraries
torch==2.3.0
transformers==4.37.2
scikit-learn==1.5.0
numpy==1.26.4

# Utilities
python-decouple==3.8

# Django Dependencies
asgiref==3.8.1
sqlparse==0.5.3
tzdata==2025.2
